repackage: moved viper configuration to a separate package
