+++ cmd/configuration/viper_test.go
@@ -1,4 +1,20 @@
-package cmd
+/*
+ * Copyright 2022 The Gremlins Authors
+ *
+ *    Licensed under the Apache License, Version 2.0 (the "License");
+ *    you may not use this file except in compliance with the License.
+ *    You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *    Unless required by applicable law or agreed to in writing, software
+ *    distributed under the License is distributed on an "AS IS" BASIS,
+ *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *    See the License for the specific language governing permissions and
+ *    limitations under the License.
+ */
+
+package configuration
 
 import (
 	"os"
@@ -68,7 +84,7 @@ func TestConfiguration(t *testing.T) {
 					os.Setenv(e.name, e.value)
 				}
 			}
-			v := getViper(tc.configPaths)
+			v := GetViper(tc.configPaths)
 
 			for key, wanted := range tc.wantedConfig {
 				got := v.Get(key)
